the convert to OpenFOAM lib is still under progress
