+++ TnbMeshPost/TnbMeshPost.vcxproj
@@ -248,6 +248,9 @@ for /r "$(ProjectDir)tnbLib" %%f in (*.hxx) do @xcopy "%%f" "$(ProjectDir)includ
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="TnbLib\Post\Convertor\OpenFOAM\2d\MeshPost2d_ConvertToOpenFOAM.cxx" />
+    <ClCompile Include="TnbLib\Post\Convertor\OpenFOAM\2d\MeshPost2d_ConvertToOpenFOAMGraph.cxx" />
+    <ClCompile Include="TnbLib\Post\Convertor\OpenFOAM\2d\MeshPost2d_ConvertToOpenFOAM_IO.cxx" />
+    <ClCompile Include="TnbLib\Post\Convertor\OpenFOAM\2d\MeshPost2d_ConvertToOpenFOAM_Mesh2d.cxx" />
     <ClCompile Include="TnbLib\Post\Quality\2d\Iso\Vlrms2Ratio\MeshPost2d_QualityMap_Vlrms2Ratio.cxx" />
     <ClCompile Include="TnbLib\Post\Quality\Surface\UniMetric\Vlrms2Ratio\MeshPost2d_QualityMapSurfaceUniMetric_Vlrms2Ratio.cxx" />
     <ClCompile Include="TnbLib\Post\Quality\Surface\Vlrms2Ratio\MeshPost2d_QualityMapSurface_Vlrms2Ratio.cxx" />
